voto , pagamento , provincia

+++ src/main/java/epicode/it/capstone_be/entities/voto/VotoService.java
@@ -2,10 +2,13 @@ package epicode.it.capstone_be.entities.voto;
 
 import epicode.it.capstone_be.auth.AppUser;
 import epicode.it.capstone_be.auth.AppUserRepository;
+import epicode.it.capstone_be.auth.Role;
 import epicode.it.capstone_be.entities.componimenti_concorso.componimento.Componimento;
 import epicode.it.capstone_be.entities.componimenti_concorso.componimento.ComponimentoRepo;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -19,30 +22,41 @@ public class VotoService {
         if(!componimentoRepo.existsById(voto.getId_componimento())) {
             throw new EntityNotFoundException("Componimento non trovato");
         }
-        if(!appUserRepo.existsById(voto.getId_user())) {
+        if(!appUserRepo.existsById(voto.getId_giudice())) {
             throw new EntityNotFoundException("Utente non trovato");
         }
         Componimento c = componimentoRepo.findById(voto.getId_componimento()).get();
-        AppUser u = appUserRepo.findById(voto.getId_user()).get();
+        AppUser u = appUserRepo.findById(voto.getId_componimento()).get();
+
         Voto v = new Voto();
         v.setComponimento(c);
         v.setUser(u);
-        v.setVoto(voto.getVoto());
+        System.out.println(voto.getVoto());
+        v.setVoto( voto.getVoto());
         return votoRepo.save(v);
     }
 
-    public void deleteVoto(Long id) {
+    public void deleteVoto(Long id, String username) {
+        AppUser u = appUserRepo.findByUsername(username).get();
+        Voto v = votoRepo.findById(id).get();
+        if(!v.getUser().equals(u)) {
+            throw new EntityNotFoundException("Non sei il proprietario di questo voto");
+        }
         if(!votoRepo.existsById(id)) {
             throw new EntityNotFoundException("Voto non trovato");
         }
         votoRepo.deleteById(id);
     }
 
-    public Voto modifyVoto(double voto, Long id) {
+    public Voto modifyVoto(Float voto, Long id, String username) {
         if(!votoRepo.existsById(id)) {
             throw new EntityNotFoundException("Voto non trovato");
         }
+        AppUser u = appUserRepo.findByUsername(username).get();
         Voto v = votoRepo.findById(id).get();
+        if(!v.getUser().equals(u)) {
+            throw new EntityNotFoundException("Non sei il proprietario di questo voto");
+        }
         if (voto < 0 || voto > 10) {
             throw new IndexOutOfBoundsException("Voto non valido") ;
         }
